requirements.txt
Django==2.2.13
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
drf-spectacular==0.22.0
psycopg2-binary==2.8.6
